requirements.txt
asgiref==3.3.1
boto3==1.17.16
botocore==1.20.16
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cssselect2==0.4.1
defusedxml==0.7.0rc2
Django==3.1.3
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-storages==1.11.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jmespath==0.10.0
livereload==2.6.3
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycairo==1.20.0
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
Pyphen==0.10.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
stripe==2.55.1
tinycss2==1.1.0
tornado==6.1
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
